move ui screen files to separate folders
