+++ flutter/lib/main.dart
@@ -6,16 +6,16 @@ import 'package:flutter/services.dart';
 import 'package:mlperfbench_common/firebase/manager.dart';
 import 'package:provider/provider.dart';
 
-import 'package:mlperfbench/app_constants.dart';
 import 'package:mlperfbench/backend/unsupported_device_exception.dart';
 import 'package:mlperfbench/benchmark/state.dart';
 import 'package:mlperfbench/build_info.dart';
-import 'package:mlperfbench/localizations/app_localizations.dart';
 import 'package:mlperfbench/resources/utils.dart';
 import 'package:mlperfbench/store.dart';
-import 'package:mlperfbench/ui/main_screen.dart';
-import 'package:mlperfbench/ui/unsupported_device_screen.dart';
+import 'package:mlperfbench/ui/root/main_screen.dart';
+import 'package:mlperfbench/ui/root/unsupported_device_screen.dart';
 import 'device_info.dart';
+import 'ui/root/app.dart';
+import 'ui/root/exception_screen.dart';
 
 // TODO sharing screen temporarily disabled
 // import 'package:mlperfbench/ui/share_screen.dart';
@@ -80,78 +80,5 @@ void autostartHandler(BenchmarkState state, Store store) async {
     print(jsonToStringIndented(result));
     print(const String.fromEnvironment('terminalStringMark'));
     exit(0);
-  }
-}
-
-class MyApp extends StatelessWidget {
-  final Widget home;
-
-  MyApp(this.home);
-
-  // This widget is the root of your application.
-  @override
-  Widget build(BuildContext context) {
-    // TODO sharing screen temporarily disabled
-    // final store = context.watch<Store>();
-
-    return MaterialApp(
-      title: 'MLPerf Mobile',
-      localizationsDelegates: [AppLocalizations.delegate],
-      supportedLocales: [const Locale('en', '')],
-      theme: ThemeData(
-        // This is the theme of your application.
-        //
-        // Try running your application with "flutter run". You'll see the
-        // application has a blue toolbar. Then, without quitting the app, try
-        // changing the primarySwatch below to Colors.green and then invoke
-        // "hot reload" (press "r" in the console where you ran "flutter run",
-        // or simply save your changes to "hot reload" in a Flutter IDE).
-        // Notice that the counter didn't reset back to zero; the application
-        // is not restarted.
-        primarySwatch: AppColors.primary,
-        // This makes the visual density adapt to the platform that you run
-        // the app on. For desktop platforms, the controls will be smaller and
-        // closer together (more dense) than on mobile platforms.
-        visualDensity: VisualDensity.adaptivePlatformDensity,
-        // This sets default color for active switcher
-        toggleableActiveColor: AppColors.secondary,
-        scaffoldBackgroundColor: AppColors.lightBackground,
-
-        // This theme of application app bar
-        appBarTheme: AppBarTheme(
-          titleTextStyle: TextStyle(color: AppColors.darkText, fontSize: 20),
-          elevation: 0,
-          backgroundColor: AppColors.lightAppBarBackground,
-          iconTheme: IconThemeData(color: AppColors.darkAppBarIconTheme),
-        ),
-      ),
-      // TODO sharing screen temporarily disabled
-      // home: store.isShareOptionChosen() ? MyHomePage() : ShareScreen(),
-      home: home,
-    );
-  }
-}
-
-class ExceptionWidget extends StatelessWidget {
-  final Object error;
-  final StackTrace stackTrace;
-
-  ExceptionWidget(this.error, this.stackTrace);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-        title: 'MLPerf Mobile',
-        home: Material(
-          type: MaterialType.transparency,
-          child: SafeArea(
-              child: Text(
-            'Error: $error\n\nStackTrace: $stackTrace',
-            style: const TextStyle(
-                fontSize: 16,
-                fontWeight: FontWeight.normal,
-                color: AppColors.lightText),
-          )),
-        ));
   }
 }
