+++ flutter/lib/ui/error_dialog.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/svg.dart';
 
 import 'package:mlperfbench/app_constants.dart';
-import 'package:mlperfbench/icons.dart';
 import 'package:mlperfbench/localizations/app_localizations.dart';
+import 'package:mlperfbench/ui/icons.dart';
 
 Future<void> showPopupDialog(BuildContext context, String header,
     SvgPicture? icon, List<String> errors) async {
