+++ flutter/lib/ui/root/unsupported_device_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
 import 'package:mlperfbench/app_constants.dart';
-import 'package:mlperfbench/icons.dart' show AppIcons;
 import 'package:mlperfbench/localizations/app_localizations.dart';
+import 'package:mlperfbench/ui/icons.dart' show AppIcons;
 import 'package:mlperfbench/ui/page_constraints.dart';
 
 class UnsupportedDeviceScreen extends StatelessWidget {
