+++ flutter/lib/ui/run/progress_screen.dart
@@ -5,7 +5,7 @@ import 'package:provider/provider.dart';
 import 'package:mlperfbench/app_constants.dart';
 import 'package:mlperfbench/benchmark/state.dart';
 import 'package:mlperfbench/localizations/app_localizations.dart';
-import 'package:mlperfbench/ui/progress_circles.dart';
+import 'package:mlperfbench/ui/run/progress_circles.dart';
 
 class ProgressScreen extends StatefulWidget {
   @override
